+++ CHANGES.md
@@ -1,9 +1,11 @@
 0.1.6
 =====
 
+* Adds uncatchable panic exception throwing using Text message.
 * Removes ``printf``
 * Removes ``string-conv`` dependency so Stack build works without ``extra-deps``.
 * Brings ``Callstack`` machinery in for GHC 8.x.
+* Removes ``throw`` and ``assert`` from ``Control.Exception`` exports.
 
 0.1.5
 =====